@@ -27,68 +25,47 @@
   },
   "homepage": "https://github.com/OTrainFans/OTrainMap#readme",
   "devDependencies": {
-    "@types/d3-ease": "3.0.0",
-    "@types/geojson": "7946.0.10",
-    "@types/mapbox-gl": "2.7.10",
-    "@types/react": "18.0.32",
-    "@types/react-custom-scroll": "4.3.2",
-    "@types/react-custom-scrollbars": "4.0.10",
-    "@types/react-dom": "18.0.11",
-    "@types/react-virtualized": "9.21.21",
-    "@types/sql.js": "1.4.4",
-    "@types/styled-components": "5.1.26",
+    "@types/geojson": "7946.0.14",
+    "@types/mapbox-gl": "2.7.21",
+    "@types/react": "18.2.64",
+    "@types/react-dom": "18.2.21",
+    "@types/react-virtualized": "9.21.29",
+    "@types/sql.js": "1.4.9",
+    "@vitejs/plugin-react": "^4.2.1",
     "ajv": "8.12.0",
     "ajv-formats": "2.1.1",
     "concurrently": "^8.0.0",
-    "copy-webpack-plugin": "11.0.0",
-    "css-loader": "6.7.3",
-    "dotenv": "16.0.3",
-    "file-loader": "6.2.0",
-    "fork-ts-checker-webpack-plugin": "8.0.0",
-    "html-webpack-plugin": "5.5.0",
-    "jest": "29.5.0",
-    "mini-css-extract-plugin": "2.7.5",
-    "nodemon": "^2.0.20",
-    "pnp-webpack-plugin": "1.7.0",
-    "raw-loader": "4.0.2",
-    "source-map-loader": "4.0.1",
-    "ts-loader": "9.4.2",
-    "typescript": "5.0.3",
-    "url-loader": "4.1.1",
-    "webpack": "5.77.0",
-    "webpack-bundle-analyzer": "4.8.0",
-    "webpack-cli": "5.0.1",
-    "webpack-dev-server": "4.13.2",
-    "webpack-merge": "5.8.0"
+    "dotenv": "16.4.5",
+    "jest": "29.7.0",
+    "nodemon": "^3.0.0",
+    "typescript": "5.4.2",
+    "vite": "^5.1.5",
+    "vite-plugin-static-copy": "^1.0.1"
   },
   "dependencies": {
-    "@mui/icons-material": "5.11.11",
-    "@mui/material": "5.11.15",
-    "@mui/styled-engine": "npm:@mui/styled-engine-sc@5.11.11",
-    "@mui/styles": "5.11.13",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "5.15.12",
+    "@mui/material": "5.15.12",
+    "@mui/styled-engine": "npm:@mui/styled-engine-sc@5.14.12",
     "@turf/bbox": "6.5.0",
     "@turf/bbox-polygon": "6.5.0",
     "@turf/boolean-contains": "6.5.0",
+    "@turf/boolean-intersects": "6.5.0",
     "@turf/boolean-overlap": "6.5.0",
     "@turf/buffer": "6.5.0",
     "bbox": "0.4.0",
-    "clsx": "1.2.1",
-    "immer": "9.0.21",
-    "mapbox-gl": "2.13.0",
+    "clsx": "2.1.0",
+    "immer": "10.0.4",
+    "mapbox-gl": "2.15.0",
     "react": "18.2.0",
-    "react-custom-scroll": "5.0.0",
-    "react-custom-scrollbars": "4.2.1",
+    "react-custom-scrollbars-2": "4.5.0",
     "react-dom": "18.2.0",
     "react-is": "18.2.0",
-    "react-map-gl": "7.0.21",
+    "react-map-gl": "7.1.7",
     "sql.js": "workspace:./sqljs",
-    "styled-components": "5.3.9",
     "tiny-warning": "1.0.3"
   },
-  "resolutions": {
-    "@mui/styled-engine": "npm:@mui/styled-engine-sc@5.11.11"
-  },
-  "packageManager": "yarn@3.5.0",
   "workspaces": [
     "./sqljs"
   ]
