@@ -1,6 +1,7 @@
 /** @type {import('@jest/types').Config.InitialOptions} */
 const config = {
   roots: ["<rootDir>/tests/"],
+  transform: {},
 };
 
-module.exports = config;
+export default config;
