@@ -1,13 +1,13 @@
-const fs = require("fs");
-const Ajv = require("ajv");
+import { readFileSync } from "fs";
+import Ajv from "ajv";
+import addFormats from "ajv-formats";
+
+import configSchema from "../config.schema.json" assert { type: "json" };
+import dataSchema from "../data.schema.json" assert { type: "json" };
+import config from "../src/config.json" assert { type: "json" };
+
 const ajv = new Ajv();
-
-const addFormats = require("ajv-formats");
 addFormats(ajv);
-
-const configSchema = require("../config.schema.json");
-const dataSchema = require("../data.schema.json");
-const config = require("../src/config.json");
 
 expect.extend({
   validatedBy(received, schema) {
@@ -33,7 +33,8 @@ test("validate config.json against config schema", () => {
 test.each(Object.values(config.agencies).flatMap((agency) => agency.data))(
   "validate %s against data schema",
   (file) => {
-    const data = require(`../data/${file}`);
+    const content = readFileSync(`./data/${file}`);
+    const data = JSON.parse(content);
     expect(data).validatedBy(dataSchema);
   }
 );
