@@ -1,4 +1,4 @@
-import React, { useRef } from "react";
+import React, { useRef, useMemo } from "react";
 import {
   Checkbox,
   Collapse,
@@ -11,7 +11,7 @@ import {
   Typography,
   useTheme,
 } from "@mui/material";
-import styled from "styled-components";
+import styled from "@emotion/styled";
 import { MenuDrawer } from "./MenuDrawer";
 import { Agency } from "../../config";
 import { LoadedMetadata } from "../../hooks/useData";
@@ -181,20 +181,24 @@ const LegendGroup = (props: LegendGroupProps) => {
   );
 };
 
-export const LegendDrawer = (props: LegendDrawerProps) => {
+export const LegendDrawer = React.memo((props: LegendDrawerProps) => {
   const { legendDrawerOpen, setLegendDrawerOpen } = useAppState();
 
-  const agencyMap: { [key: string]: LoadedMetadata[] } = {};
-  Object.values(
-    metadata as unknown as { [key: string]: LoadedMetadata }
-  ).forEach((line) => {
-    const agency = line.agency;
-    if (agencyMap[agency] == null) {
-      agencyMap[agency] = [];
-    }
+  const agencyMap: { [key: string]: LoadedMetadata[] } = useMemo(() => {
+    const map: { [key: string]: LoadedMetadata[] } = {};
+    Object.values(
+      metadata as unknown as { [key: string]: LoadedMetadata }
+    ).forEach((line) => {
+      const agency = line.agency;
+      if (map[agency] == null) {
+        map[agency] = [];
+      }
 
-    agencyMap[agency].push(line);
-  });
+      map[agency].push(line);
+    });
+
+    return map;
+  }, []);
 
   return (
     <MenuDrawer
@@ -230,4 +234,4 @@ export const LegendDrawer = (props: LegendDrawerProps) => {
       </List>
     </MenuDrawer>
   );
-};
+});
